+++ requirements.txt
@@ -1,0 +1,10 @@
+torch
+torchvision
+torchaudio
+transformers
+bitsandbytes
+accelerate
+datasets
+pandas 
+trl
+nltk
